Smove cron Schedules, divided weeklyReport logic
