+++ src/routes/messages/messages.api.ts
@@ -5,12 +5,10 @@ import auth from '../../middleware/auth';
 import { Message } from '../../models/message.model';
 import { Outcome } from '../../models/outcome.model';
 import { Patient } from '../../models/patient.model';
-import runCronSchedules from './cronSchedules';
+import runCronSchedules from '../../background_jobs/cronSchedules';
 
 import initializeScheduler from '../../utils/scheduling';
 import errorHandler from '../error';
-
-
 
 const router = express.Router();
 initializeScheduler();
